orcun/requirements.txt
tensorflow==2.15.0
uvicorn==0.27.1
pillow==10.2.0
requests==2.31.0
fastapi==0.109.2
python-multipart==0.0.9
torch==2.2.0
torchvision==0.17.0